==> Part1- DataCleaning&FeatureEngineering.ipynb <==
# UK Road Safety: Traffic Accidents and Vehicles

## PART - 1 - Data Cleaning and Feature Engineering

 Dataset Link - https://www.kaggle.com/datasets/tsiaras/uk-road-safety-accidents-and-vehicles

### Uploading Data

import pandas as pd
import numpy as np 
import matplotlib.pyplot as plt
import seaborn as sns

# read a csv file 
df = pd.read_csv('Accident_Information.csv', header=0, encoding='unicode_escape', low_memory=False)


### Interpreting and understanding the data

# Display the first 10 rows of the dataframe
df.head(10)

df.shape

df.describe() #describe the data

df.dtypes #check data types of each column

# Identifying variables with object datatype
df.dtypes[df.dtypes == 'object']

#typecasting object data type feature to category type

df['Accident_Severity'] = df['Accident_Severity'].astype('category')
df['Day_of_Week'] = df['Day_of_Week'].astype('category')
df['Road_Type'] = df['Road_Type'].astype('category')
df['Weather_Conditions'] = df['Weather_Conditions'].astype('category')
df['Road_Surface_Conditions'] = df['Road_Surface_Conditions'].astype('category')
df['Light_Conditions'] = df['Light_Conditions'].astype('category')
df['Special_Conditions_at_Site'] = df['Special_Conditions_at_Site'].astype('category')
df['Junction_Control'] = df['Junction_Control'].astype('category')
df['Junction_Detail'] = df['Junction_Detail'].astype('category')


# Check the new data type
print(df['Accident_Severity'].dtype)
print(df['Day_of_Week'].dtype)
print(df['Road_Type'].dtype)
print(df['Weather_Conditions'].dtype)
print(df['Road_Surface_Conditions'].dtype)
print(df['Light_Conditions'].dtype)
print(df['Special_Conditions_at_Site'].dtype)
print(df['Junction_Control'].dtype)
print(df['Junction_Detail'].dtype)



#typecasting object data type feature to datetime data type
df['Date'] = pd.to_datetime(df['Date'])



df['Time'] = pd.to_datetime(df['Time']).dt.time

# Check the new data type
print(df['Date'].dtype)
print(df['Time'].dtype)

# Calculate total number of cells in dataframe
df.columns
df.index
print("Total number of columns are: ", len(df.columns))
print("Total number of rows are: ", len(df.index))
print("Total number of cells in dataframe: ", len(df.columns) * len(df.index))


df.columns

Array1 = [ABC for ABC in df.columns if df[ABC].dtypes !='O']
print(Array1)
print(len(Array1))

# Count of numerical features 
numerical_features = [f for f in df.columns if df[f].dtypes!='O']
print('Count of Numerical Features :',len(numerical_features))


# Count of Categorical features 
cat_features = [c for c in df.columns if df[c].dtypes=='O']
print('Count of Categorical Features :',len(cat_features))

#find duplicates and return 0 if there aren't any duplicates
df.duplicated().sum() 

#Checking for null values
df.isnull().sum() #returns the number of missing values in the dataset.

## Value Counts

#Return a Series containing counts of unique values
df['1st_Road_Class'].value_counts() 

df['1st_Road_Number'].value_counts() 

df['Accident_Severity'].value_counts() 

df['Carriageway_Hazards'].value_counts() 

df['Date'].value_counts() 

df['Day_of_Week'].value_counts() 

df['Did_Police_Officer_Attend_Scene_of_Accident'].value_counts() 

df['Junction_Control'].value_counts() 

df['Junction_Detail'].value_counts() 

df['Latitude'].value_counts() 

df['Light_Conditions'].value_counts() 

df['Longitude'].value_counts() 

df['LSOA_of_Accident_Location'].value_counts() 

df['Pedestrian_Crossing-Human_Control'].value_counts() 

df['Pedestrian_Crossing-Physical_Facilities'].value_counts() 

df['Road_Type'].value_counts() 

df['Special_Conditions_at_Site'].value_counts() 

df['Speed_limit'].value_counts() 

df['Time'].value_counts() 

df['Urban_or_Rural_Area'].value_counts() 

df['Weather_Conditions'].value_counts() 

df['Year'].value_counts() 

df['InScotland'].value_counts() 

## Correlation of the features 

df.corr()

## Data Cleaning and Feature Engineering


df.isnull().sum()

#calulate mean latitude for missing value of latitude(use local authority as it provides good approximation of the area where the accidents occured)
df1 = df.groupby('Local_Authority_(District)')['Latitude'].mean()
df['Latitude'] = df.apply(lambda x: df1[x['Local_Authority_(District)']] if pd.isna(x['Latitude']) else x['Latitude'], axis=1)

#refrence from-https://stackoverflow.com/questions/18689823/pandas-dataframe-replace-nan-values-with-average-of-columns

#calulate mean longitude for missing value of longitude(use local authority as it provides good approximation of the area where the accidents occured)
df1 = df.groupby('Local_Authority_(District)')['Longitude'].mean()
df['Longitude'] = df.apply(lambda x: df1[x['Local_Authority_(District)']] if pd.isna(x['Longitude']) else x['Longitude'], axis=1)

#dropping unecessary columns
dropped_col=['Accident_Index','Location_Easting_OSGR','Location_Northing_OSGR','Police_Force','Local_Authority_(District)','Local_Authority_(Highway)']
df.drop(dropped_col,axis=1,inplace=True)

df = df.drop('2nd_Road_Class', axis=1)

df[df['1st_Road_Number'].isnull()]

#removing row with null values
df = df.dropna(subset=['1st_Road_Number'])

df = df.drop('2nd_Road_Number', axis=1)

# Count missing values by category

data=df[df['Did_Police_Officer_Attend_Scene_of_Accident'].isnull()]
data

print(df['Did_Police_Officer_Attend_Scene_of_Accident'].unique())

#using mode to fill the nan values with the mode of the non missing values.
df1 = df['Did_Police_Officer_Attend_Scene_of_Accident'].mode()[0]
df['Did_Police_Officer_Attend_Scene_of_Accident'].fillna(df1, inplace=True)


data=df[df['Did_Police_Officer_Attend_Scene_of_Accident'].isnull()]
data

# Remove rows with missing values in the 'LSOA_of_Accident_Location' column
df = df.dropna(subset=['LSOA_of_Accident_Location'])

#Fill missing Pedestrian_Crossing-Human_Control values with the mode
df1 = df['Pedestrian_Crossing-Human_Control'].mode()[0]
df['Pedestrian_Crossing-Human_Control'].fillna(df1, inplace=True)

#Fill missing Pedestrian_Crossing-Physical_Facilities values with the mode
df1 = df['Pedestrian_Crossing-Physical_Facilities'].mode()[0]
df['Pedestrian_Crossing-Physical_Facilities'].fillna(df1, inplace=True)


#fill the missing values in the 'Speed limit' column with corresponding median speed limit for the road type
df1 = df.groupby('Road_Type')['Speed_limit'].median()
df['Speed_limit'] = df.apply(lambda x: df1[x['Road_Type']] if pd.isna(x['Speed_limit']) else x['Speed_limit'], axis=1)

#fill missing value in 'Time' with mode
df1 = df['Time'].mode()[0]
df['Time'].fillna(df1, inplace=True)

df.isnull().sum()

df.columns

## Exporting the Cleaned Dataset as CSV for next steps 

df.shape

# saving the dataframe to csv 
df.to_csv('UK Road Safety cleaned data.csv',index=False)

# Read saved file and display  top 5 rows 
df_new = pd.read_csv('UK Road Safety cleaned data.csv')
df_new.head(n=5)
