==> Part2-Dashboard-TIme Series Analysis.ipynb <==
# UK Road Safety Data Set

## Part 2 - Dashboard

## Uploading the data after cleaning and feature engineering

import pandas as pd
import numpy as np 
import matplotlib.pyplot as plt
import seaborn as sns

df= pd.read_csv('UK Road Safety cleaned data.csv')

df.columns

df.describe()

###  Road Accidents in the UK over specific time period.

import plotly.express as px
import dash
import dash_core_components as dcc
import dash_html_components as html
from dash.dependencies import Input, Output

# Total accidents by year

# Convert 'Date' and 'Time' columns to datetime format
df["Date"] = pd.to_datetime(df["Date"])
df["Time"] = pd.to_datetime(df["Time"], format="%H:%M:%S")

# Extract year from 'Date' column and create new 'Year' column
df["Year"] = df["Date"].dt.year

# Create charts
accidents_by_year_linegraph = px.line(df.groupby("Year").size().reset_index(name="Total Accidents"), x="Year", y="Total Accidents", title="Total Accidents by Year")

severity_counts_pie = px.pie(df, names="Accident_Severity", title="Accident Severity Counts")

urban_vs_rural_histo = px.histogram(df, x="Urban_or_Rural_Area", color="Accident_Severity", title="Accidents by Urban vs Rural Area")


month_of_year_bargraph = px.bar(df.groupby(df["Date"].dt.strftime("%B")).size().reset_index(name="Count"), x="Date", y="Count", title="Accidents by Month")


# Create pivot table with accidents by month and day of week
df_pivot = pd.pivot_table(df, values='Accident_Severity', index=df['Date'].dt.month_name(), columns=df['Day_of_Week'], aggfunc='count')

# Create pivot table with accidents by hour of day and day of week
df_pivot_hour = pd.pivot_table(df, values='Accident_Severity', index=df['Date'].dt.hour, columns=df['Day_of_Week'], aggfunc='count')

accidents_by_hour_bargraph = px.bar(df_pivot_hour, barmode="group", title="Accidents by Hour of Day and Day of Week")




# Create app layout
app = dash.Dash(__name__)
app.layout = html.Div(children=[
    dcc.RangeSlider(
        id='year-slider',
        min=df['Year'].min(),
        max=df['Year'].max(),
        value=[2005, 2015],
        marks={str(year): str(year) for year in df['Year'].unique()},
        step=None
    ),
    html.Br(),
    html.H1(children="UK Road Accidents Dashboard"),
    dcc.Graph(id="line-chart", figure=accidents_by_year_linegraph),
    dcc.Graph(id="pie-chart", figure=severity_counts_pie),
    dcc.Graph(id="histogram", figure=urban_vs_rural_histo),
    dcc.Graph(id="month-of-year-bargraph", figure=month_of_year_bargraph),
    dcc.Graph(id="pivot-table", figure=px.imshow(df_pivot, x=df_pivot.columns, y=df_pivot.index, title="Accidents by Month and Day of Week")),
    dcc.Graph(id="bar-chart-hour", figure=accidents_by_hour_bargraph)
])

# Define callback to update graphs
@app.callback(
    [Output("line-chart", "figure"), Output("pie-chart", "figure"), Output("histogram", "figure"), Output("month-of-year-bargraph", "figure"), Output("pivot-table", "figure"), Output("bar-chart-hour", "figure")],
    [Input("year-slider", "value")]
)
def update_graphs(year_range):
    filtered_df = df[(df["Year"] >= year_range[0]) & (df["Year"] <= year_range[1])]
    
    updated_accidents_by_year = px.line(filtered_df.groupby("Year").size().reset_index(name="Total Accidents"), x="Year", y="Total Accidents", title="Total Accidents by Year")
    
    updated_severity_counts = px.pie(filtered_df, names="Accident_Severity", title="Accident Severity Counts")
    
    updated_urban_vs_rural = px.histogram(filtered_df, x="Urban_or_Rural_Area", color="Accident_Severity", title="Accidents by Urban vs Rural Area")
    

    updated_month_of_year = px.bar(filtered_df.groupby(filtered_df["Date"].dt.strftime("%B")).size().reset_index(name="Count"), x="Date", y="Count", title="Accidents by Month")

    
    df_pivot = pd.pivot_table(filtered_df, values='Accident_Severity', index=filtered_df['Date'].dt.month_name(), 
                              columns=filtered_df['Day_of_Week'], aggfunc='count')
    
    df_pivot_hour = pd.pivot_table(filtered_df, values='Accident_Severity', index=filtered_df['Date'].dt.hour, columns=filtered_df['Day_of_Week'], aggfunc='count')
    
    updated_accidents_by_month_day = px.imshow(df_pivot, x=df_pivot.columns, y=df_pivot.index, 
                                                title="Accidents by Month and Day of Week")
    updated_accidents_by_hour = px.bar(df_pivot_hour, barmode="group", title="Accidents by Hour of Day and Day of Week")
    
    return updated_accidents_by_year, updated_severity_counts, updated_urban_vs_rural, updated_month_of_year, updated_accidents_by_month_day, updated_accidents_by_hour


# Run app
if __name__ == "__main__":
    app.run_server()
    
#note when using slider, it takes time to get the output(min 2 minutes).

This dashboard summarizes the information about road accidents in the UK.I have created 4 graph which will help us visualize the information.The dashboard includes four charts that display different aspects of the data:
- A line chart showing the total number of accidents by year.
- A pie chart showing the distribution of accidents by severity level.
- A histogram showing the distribution of accidents by urban or rural area, with different severity levels represented by different colors(BLUE-Slight,RED-Serious,GREEN-Fatal).
- A bar chart showing the number of accidents by day of the week.
- A heatmap of accidents by month and day of the week.
- A grouped bar chart of accidents by hour of the day and day of the week.
