==> Part3-simple regression.ipynb <==
import pandas as pd
from sklearn.linear_model import LinearRegression
from sklearn.metrics import r2_score
import numpy as np
import logging

df= pd.read_csv('UK Road Safety cleaned data.csv')

df.columns

# Select features and target variable
X = df[['Speed_limit']]
y = df['Number_of_Vehicles']

# Split data into train, test, and new data
train_size = 0.7
test_size = 0.2
new_size = 0.1

train_end_idx = int(len(df)*train_size)
test_end_idx = train_end_idx + int(len(df)*test_size)

X_train, y_train = X[:train_end_idx], y[:train_end_idx]
X_test, y_test = X[train_end_idx:test_end_idx], y[train_end_idx:test_end_idx]
X_new, y_new = X[test_end_idx:], y[test_end_idx:]

# Train model
model = LinearRegression()
model.fit(X_train, y_train)


# Make predictions on train and test data
y_train_pred = model.predict(X_train)
y_test_pred = model.predict(X_test)


# Evaluate model accuracy
train_r_squared = r2_score(y_train, y_train_pred)
test_r_squared = r2_score(y_test, y_test_pred)
print(f"Train R-squared value for the model: {train_r_squared:.2f}")
print(f"Test R-squared value for the model: {test_r_squared:.2f}")


# Write log file
logging.basicConfig(filename='regression_log.txt', level=logging.INFO)
logging.info(f"Train R-squared value for the model: {train_r_squared:.2f}")
logging.info(f"Test R-squared value for the model: {test_r_squared:.2f}")

# Score new data
y_new_pred = model.predict(X_new)

# Round predictions to nearest whole number
y_test_pred = np.round(y_test_pred)
y_new_pred = np.round(y_new_pred)

# Generate prediction file
prediction_df = pd.DataFrame({'Speed Limit': X_new['Speed_limit'], 'Actual Number of Vehicles': y_new, 'Predicted Number of Vehicles': y_new_pred})
prediction_df.to_csv('prediction.csv', index=False)
